Add Extract Text to the menu
I finally figured out how to add a command to the menu, which is by
editing the file OneMore/Ribbon/Ribbon.xml.

There is no icon in MSO for OCR, but the export text icon is pretty darn
close, so I am using it instead.  To find a list of icons search for
imageMSO gallery.