Start a unit test project
I find testing in OneMore to be hard.  I am not familiar with the
programming model or UI programming.  My iteration loop is build, build
installer, install OneMore, start OneNote, attach debugger, find bug,
and rinse and repeat.  It takes minutes at the best.  I tried to debug
why Extract Text was not visible in the OneMore menu, and after several
cycles of rinse and repeat I resorted to a unit test (should have done
it earlier).  The unit test proved that it was a valid menu option, and
helped me to look elsehwere.  The bug was not adding it to Ribbon.xml.

There is one test in the project, and asserts little.  The test is
useful for debugging and understanding context without having to install
the plugin.